requirements.txt
amqp==5.0.9
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.1
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
idna==3.3
kombu==5.2.3
packaging==21.3
prompt-toolkit==3.0.24
psycopg2==2.9.3
pyparsing==3.0.6
pytz==2021.3
redis==4.1.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
